Split Rummager between advanced search and search
Finders in Whitehall used the '/advanced_search' endpoint in Rummager. This
does not surface all relevant content because it only uses the goverment index.

This change makes the announcement finder use the newer /search endpoint that
spans the government and govuk search indices and should therefore be capable
of returning all relevant results.

Since we only need to make the change to announcements at this stage, we split
out the Rummager document filter between advanced search and search so that the
right queries are constructed for each endpoint.