Split Document Filter Result
Now that we are using the '/search' endpoint for the Announcements filter page,
we will need to process the Rummager results in a different way to those that
are returned from '/advanced-search'.  This splits the Document Filter::Result
class into Document Filter::AdvancedSearchResult and Document
Filter::SearchResult which will process the Rummager results in the appropriate
way.