Display results from govuk index in atom feed
Changing the search index we call in Rummager from `government` to
`govuk` meant that the results weren’t displaying in the atom feed.

The title, author and updated date were being displayed, but all of the
“entries” (the results of the search) were missing.

This is because the `government` index was also returning the edition id
in the results, and the controller was then using the id to get
Announcements from the database rather than using the results returned
from Rummager.

Use RummagerDocumentPresenter introduced in PR #4457 to format the
results from Rummager into state that can be read by the atom feed
builder.

Consequences for the user:
1. The id of each entry in the feed will change because we are no longer
using the edition id

2. The entries will no longer display the body of the content as
Rummager doesn't store this information