Include has_brexit_update flag in applicable document details
This adds the has_brexit_flag to the details in the content we send
to the publishing-api for various document types.

NOTE: This commit is incomplete. There may be other documents in
Whitehall that we want to send to the publishing api with this
flag. Need to check with content designers / product manager.

Additionally, a change to govuk-content-schemas is required and
has not yet been done.